+++ app/src/androidTest/java/com/flopp/razdor/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package good.damn.ui
+package com.flopp.razdor
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("good.flopp.ui.test", appContext.packageName)
+        assertEquals("com.flopp.floop.flopp.razdor", appContext.packageName)
     }
 }
