--- app/src/main/java/com/example/zov_android/extensions/okhttp/RequestBuilder.kt
@@ -1,7 +1,0 @@
-package com.example.zov_android.extensions.okhttp
-
-import okhttp3.ResponseBody
-import org.json.JSONObject
-
-fun ResponseBody.toJSON() =
-    JSONObject(string())
