--- app/src/main/java/com/example/zov_android/extensions/view/AppCompatEditText.kt
@@ -1,14 +1,0 @@
-package com.example.zov_android.extensions.view
-
-import androidx.appcompat.widget.AppCompatEditText
-import com.example.zov_android.EZApp
-
-fun AppCompatEditText.applyTheme() = EZApp.theme.apply {
-    setHintTextColor(
-        colorHint
-    )
-
-    setTextColor(
-        colorText
-    )
-}
