--- app/src/main/java/com/example/zov_android/model/EZModelUser.kt
@@ -1,7 +1,0 @@
-package com.example.zov_android.model
-
-data class EZModelUser(
-    val id: String,
-    val token: String,
-    val username: String
-)
