--- app/src/main/java/com/example/zov_android/network/clients/base/EZClientBase.kt
@@ -1,8 +1,0 @@
-package com.example.zov_android.network.clients.base
-
-import androidx.annotation.StringRes
-
-interface EZClientBase {
-    val id: String
-    val username: String?
-}
