--- app/src/main/java/com/example/zov_android/themes/UIThemeDark.java
@@ -1,14 +1,0 @@
-package com.example.zov_android.themes;
-
-import good.damn.ui.theme.UITheme;
-
-public final class UIThemeDark
-extends UITheme {
-    public UIThemeDark() {
-        colorButton = 0xffffffff;
-        colorButtonText = 0xff000315;
-        colorText = 0xffffffff;
-        colorBackground = 0xff000315;
-        colorHint = 0xffffffff;
-    }
-}
