+++ app/src/main/java/com/flopp/razdor/extensions/Int.kt
@@ -1,0 +1,9 @@
+package com.flopp.razdor.extensions
+
+fun Int.convertToHumanTime() : String{
+    val seconds = this%60
+    val minutes = this/60
+    val secondsString = if (seconds<10) "0$seconds" else "$seconds"
+    val minutesString = if (minutes < 10) "0$minutes" else "$minutes"
+    return "$minutesString:$secondsString"
+}
