+++ app/src/main/java/com/flopp/razdor/extensions/okhttp/RequestBuilder.kt
@@ -1,0 +1,7 @@
+package com.flopp.razdor.extensions.okhttp
+
+import okhttp3.ResponseBody
+import org.json.JSONObject
+
+fun ResponseBody.toJSON() =
+    JSONObject(string())
