+++ app/src/main/java/com/flopp/razdor/model/EZModelUser.kt
@@ -1,0 +1,7 @@
+package com.flopp.razdor.model
+
+data class EZModelUser(
+    val id: String,
+    val token: String,
+    val username: String
+)
