+++ app/src/main/java/com/flopp/razdor/network/clients/base/EZClientBase.kt
@@ -1,0 +1,6 @@
+package com.flopp.razdor.network.clients.base
+
+interface EZClientBase {
+    val id: String
+    val username: String?
+}
