+++ app/src/main/java/com/flopp/razdor/network/clients/http/listeners/EZIObserverHttpOnAuth.kt
@@ -1,0 +1,14 @@
+package com.flopp.razdor.network.clients.http.listeners
+
+import com.flopp.razdor.enums.EZEnumStateAuth
+import com.flopp.razdor.model.EZModelUser
+
+interface EZIObserverHttpOnAuth {
+    fun onAuthSuccess(
+        user: EZModelUser
+    )
+
+    fun onAuthFailed(
+        state: EZEnumStateAuth
+    )
+}
