+++ app/src/main/java/com/flopp/razdor/themes/UIThemeDark.java
@@ -1,0 +1,14 @@
+package com.flopp.razdor.themes;
+
+import good.damn.ui.theme.UITheme;
+
+public final class UIThemeDark
+extends UITheme {
+    public UIThemeDark() {
+        colorButton = 0xffffffff;
+        colorButtonText = 0xff000315;
+        colorText = 0xffffffff;
+        colorBackground = 0xff000315;
+        colorHint = 0xffffffff;
+    }
+}
